+++ frontend/src/app/shared/regaxes/user-regax-patterns.ts
